wip: rename css files to have module suffix and fix styles
