--- tsconfig.cjs.json
@@ -1,11 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "noEmit": false,
-    "declaration": true,
-    "sourceMap": true,
-    "module": "CommonJS",
-    "outDir": "lib/cjs"
-  },
-  "files": ["src/index.ts"]
-}
