--- tsconfig.types.json
@@ -1,11 +1,0 @@
-{
-	"extends": "./tsconfig.json",
-	"files": ["src/index.ts"],
-	"compilerOptions": {
-		"noEmit": false,
-		"emitDeclarationOnly": true,
-		"declaration": true,
-		"outDir": "types"
-	},
-	"include": ["src"]
-}
